Redesigned ExchangeRate; created Repository
The ExchangeRate model was redesigned to be a value object in order to
remove its responsibility of measuring the rate. The
ExchangeRateRepository should be responsible for fetching this (since
it's an external value), and it should return the immutable and
comparable value object ExchangeRate.

One issue found, however, was that for implementing the
UnitOfAccount.valueIn(other) function in CurrencyUnit an
ExchangeRateRepository would have to be available to every CurrencyUnit
directly. Due to time concerns, valueIn was removed from UnitOfAccount,
but this is a serious nuisance because UnitOfAccount can no longer
implement Comparable<T>, and this conflicts with the very definition of
a unit of account.