Added missing Response model to Android app
The networking of the Android app was mistakenly designed to receive the
domain models directly from the JSON response, which is not the case
with the :api:application. Thus, a response model was created so that
Jackson can deserialize it, and the Repository can map it to the domain
model.

Other changes:
 - Made the sole exposed function in ViewModel launch a coroutine, so
 that all processing from this point forward is run on background
 threads. Furthermore, by relying on LiveData objects, the View will
 always do its own processing on the Main thread.